requirements.txt
fastapi
uvicorn
boto3
ffmpeg-python
openai
pymongo
python-dotenv
pydantic[email]
